Move non-Spezi components
